requirements.txt
astroid==2.2.5
Django==2.2.4
django-templated-mail==1.1.1
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
postgres==2.2.2
psycopg2-binary==2.8.3
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
